Rename hashlib to hashutil
