+++ pkg/hashutil/common_test.go
@@ -1,4 +1,4 @@
-package hashlib
+package hashutil
 
 import (
 	"crypto/rand"
